refactor: 날씨 관련 객체 weather 패키지로 이전
