+++ src/main/java/dandi/dandi/pushnotification/application/sevice/WeatherPushNotificationScheduler.java
@@ -3,11 +3,11 @@ package dandi.dandi.pushnotification.application.sevice;
 import dandi.dandi.member.application.port.out.MemberPersistencePort;
 import dandi.dandi.member.domain.Location;
 import dandi.dandi.pushnotification.application.port.out.persistence.PushNotificationPersistencePort;
-import dandi.dandi.pushnotification.application.port.out.weather.WeatherForecast;
-import dandi.dandi.pushnotification.application.port.out.weather.WeatherForecastInfoManager;
 import dandi.dandi.pushnotification.application.port.out.webpush.WebPushManager;
 import dandi.dandi.pushnotification.application.sevice.message.WeatherPushNotificationMessageGenerator;
 import dandi.dandi.pushnotification.domain.PushNotification;
+import dandi.dandi.weather.application.port.out.WeatherForecast;
+import dandi.dandi.weather.application.port.out.WeatherForecastInfoManager;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
