+++ src/main/java/dandi/dandi/pushnotification/application/sevice/message/WeatherPushNotificationMessageGeneratorImpl.java
@@ -1,6 +1,6 @@
 package dandi.dandi.pushnotification.application.sevice.message;
 
-import dandi.dandi.pushnotification.application.port.out.weather.WeatherForecast;
+import dandi.dandi.weather.application.port.out.WeatherForecast;
 import org.springframework.stereotype.Component;
 
 @Component
