+++ src/test/java/dandi/dandi/pushnotification/application/service/WeatherPushNotificationSchedulerTest.java
@@ -1,4 +1,4 @@
-package dandi.dandi.pushnotification.application.sevice;
+package dandi.dandi.pushnotification.application.service;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.any;
@@ -11,12 +11,13 @@ import static org.mockito.Mockito.when;
 import dandi.dandi.member.application.port.out.MemberPersistencePort;
 import dandi.dandi.member.domain.Location;
 import dandi.dandi.pushnotification.application.port.out.persistence.PushNotificationPersistencePort;
-import dandi.dandi.pushnotification.application.port.out.weather.WeatherForecast;
-import dandi.dandi.pushnotification.application.port.out.weather.WeatherForecastInfoManager;
 import dandi.dandi.pushnotification.application.port.out.webpush.WebPushManager;
+import dandi.dandi.pushnotification.application.sevice.WeatherPushNotificationScheduler;
 import dandi.dandi.pushnotification.application.sevice.message.WeatherPushNotificationMessageGenerator;
 import dandi.dandi.pushnotification.domain.PushNotification;
 import dandi.dandi.pushnotification.domain.PushNotificationTime;
+import dandi.dandi.weather.application.port.out.WeatherForecast;
+import dandi.dandi.weather.application.port.out.WeatherForecastInfoManager;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
